Added a unit test case for `AMbytes()`. Added a unit test case for `AMstr()`. Added a unit test case for `AMstrcmp()`. Added a unit test case for `AMstrdup()`. Added unit test cases for the generated `AM...ToString()` functions. Added unit test cases for the generated `AM...FromString()` functions. Fixed `AMstrdup()`. Ported `AMstrcmp()` from C to Rust so that it would be in the core bindings. Made the result of calling `AMbytes(NULL, ...)` safe. Added the Cargo-built core bindings to the install target. Added descriptions for the utility headers. Fixed a typo in the documentation for `AMsyncStateEncode()`.

+++ rust/automerge-c/src/actor_id.rs
@@ -1,4 +1,5 @@
 use automerge as am;
+use libc::c_int;
 use std::cell::RefCell;
 use std::cmp::Ordering;
 use std::str::FromStr;
@@ -93,7 +94,7 @@ pub unsafe extern "C" fn AMactorIdBytes(actor_id: *const AMactorId) -> AMbyteSpa
 pub unsafe extern "C" fn AMactorIdCmp(
     actor_id1: *const AMactorId,
     actor_id2: *const AMactorId,
-) -> isize {
+) -> c_int {
     match (actor_id1.as_ref(), actor_id2.as_ref()) {
         (Some(actor_id1), Some(actor_id2)) => match actor_id1.as_ref().cmp(actor_id2.as_ref()) {
             Ordering::Less => -1,
